+++ src/IdentifierResolvers/SnowflakeIdentifierResolver.php
@@ -10,8 +10,8 @@ class SnowflakeIdentifierResolver implements IdentifierResolver
 {
     public function identifier(int $time, int $sequence, string|null $group = null): int
     {
-        $workerIdLeftShift = Snowflake::SEQUENCE_BITS;
-        $datacenterIdLeftShift = Snowflake::WORKER_ID_BITS + Snowflake::SEQUENCE_BITS;
+        $workerIdLeftShift = Snowflake::sequenceBits();
+        $datacenterIdLeftShift = Snowflake::workerIdBits() + Snowflake::sequenceBits();
         $timestampLeftShift = Snowflake::ID_BITS - Snowflake::TIMESTAMP_BITS;
 
         return ($time << $timestampLeftShift)